support reading args from config file
